requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
Django==4.2.7
django-redis==5.4.0
djangorestframework==3.14.0
pika==1.1.0
psycopg2-binary==2.9.9
python-json-logger==2.0.1
pytz==2023.3.post1
redis==4.6.0
sqlparse==0.4.4
typing_extensions==4.8.0
vine==5.1.0
